Add authentication client, service, dto. Add jwt handling
Add authentication with backend by using jwt. There are many
changes so the JWT can be stored in VaadinSession and send
with requests to the backend as a header.
Class name changes CarService to CarRepair, AvailableCarService
to AvailableCarRepair.
JWT added to UserClient class, next will add it to other client
classes.